Add: Integrated ResponseEntity

+++ src/main/java/net/biswajit/journalApp/controller/JournalEntryController.java
@@ -4,6 +4,8 @@ import net.biswajit.journalApp.entity.JournalEntry;
 import net.biswajit.journalApp.service.JournalEntryService;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDate;
@@ -17,37 +19,60 @@ public class JournalEntryController {
     private JournalEntryService journalEntryService;
 
     @GetMapping
-    public List<JournalEntry> getAllEntry(){
-        return journalEntryService.getAllEntry();
+    public ResponseEntity<?> getAllEntry(){
+        List<JournalEntry> journalEntries = journalEntryService.getAllEntry();
+        if(journalEntries != null && !journalEntries.isEmpty()){
+            return new ResponseEntity<>(journalEntries, HttpStatus.OK);
+
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @GetMapping("/id/{myId}")
-    public JournalEntry getEntryById(@PathVariable ObjectId myId){
-        return journalEntryService.getEntryById(myId).orElse(null);
+    public ResponseEntity<?> getEntryById(@PathVariable ObjectId myId){
+        Optional<JournalEntry> journalEntry = journalEntryService.getEntryById(myId);
+        if(journalEntry.isPresent()){
+            return new ResponseEntity<>(journalEntry.get(),HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PostMapping
-    public JournalEntry postEntry(@RequestBody JournalEntry myEntry){
-        myEntry.setDate(LocalDate.now());
-        journalEntryService.saveEntry(myEntry);
-        return myEntry;
+    public ResponseEntity<JournalEntry> postEntry(@RequestBody JournalEntry myEntry){
+        try {
+            myEntry.setDate(LocalDate.now());
+            journalEntryService.saveEntry(myEntry);
+            return new ResponseEntity<>(myEntry,HttpStatus.CREATED);
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
     }
 
     @PutMapping("id/{myId}")
-    public boolean updateEntry(@PathVariable ObjectId myId, @RequestBody JournalEntry newEntry){
-        JournalEntry oldEntry = getEntryById(myId);
+    public ResponseEntity<?> updateEntry(@PathVariable ObjectId myId, @RequestBody JournalEntry newEntry){
+        JournalEntry oldEntry = journalEntryService.getEntryById(myId).orElse(null);
         if(oldEntry != null){
             oldEntry.setTitle(newEntry.getTitle() != null && !newEntry.getTitle().isEmpty() ? newEntry.getTitle() : oldEntry.getTitle());
             oldEntry.setContent(newEntry.getContent() != null && !newEntry.getContent().isEmpty() ? newEntry.getContent() : oldEntry.getContent());
+            journalEntryService.saveEntry(oldEntry);
+            return new ResponseEntity<>(oldEntry,HttpStatus.OK);
         }
-        journalEntryService.saveEntry(oldEntry);
-        return true;
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping("/id/{myId}")
-    public boolean deleteEntry(@PathVariable ObjectId myId){
-        journalEntryService.deleteEntryById(myId);
-        return true;
+    public ResponseEntity<?> deleteEntry(@PathVariable ObjectId myId){
+        Optional<JournalEntry> journalEntry = journalEntryService.getEntryById(myId);
+        if(journalEntry.isPresent()) {
+            journalEntryService.deleteEntryById(myId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+
+        return new ResponseEntity<>("Failed",HttpStatus.NOT_FOUND);
+
     }
 
 }
